Split test methods out of the Test class.
