+++ src/index.js
@@ -35,6 +35,7 @@ const server = http.createServer((request, response) => {
     case '/test/search.js':
     case '/view/search.js':
     case '/test.js':
+    case '/test/test.js':
     case '/view/test.js':
         response.statusCode = 200;
         response.setHeader('Content-Type', 'text/javascript');
